mlProject/requirements.txt
torch~=2.8.0
torchvision~=0.23.0
pillow~=11.3.0
numpy~=2.2.6
requests~=2.32.5
ultralytics~=8.3.203
opencv-python~=4.12.0.88
uvicorn~=0.37.0
fastapi~=0.117.1